Logger for debugging
Use a logger with level of `:info` by default, but log a few
specific events at the `:debug` level for aide in tracing the
scheduling and execution of tasks.

Each task can have it's own logger name, which can be provided
by the `:name` argument to `every`, but also set to `false` to
not have it's own name.